requirements.txt
web3==6.*
python-dotenv==1.*
py-solc-x==2.*
eth-account==0.*
eth-tester>=0.9.0b0
